Source code, binaries and examples
The source code, the generated binaries and some examples have been added to this repository. Currently more elements have been added to those published in the article, however, the examples contemplate the elements that are going to be published, so that someone who follows the instructions described in the binaries/README.md file will be able to replicate the reported results.